Removed functionality to create non-essential output
I tried to remove all functionality creating non-essential output,
i.e. output which isn't part of the local JS app or input for the
clld app.

Whatever statistics are deemed necessary for the end-user should be
computed on-the-fly from the database.

It also turned out that the "size" of a wordlist (i.e. the number of
words) isn't really used in the visualization, so doesn't need to
be computed. This makes computing languages trivial, so the
corresponding command was removed. The command `concepts` was also only
used to compute summary statistics on concepts, thus was removed.

Overall, I think, we are now approaching a codebase where adding tests
makes sense, because a decent coverage is in reach :)